Add ProfileFriendlist and ProfileSuggestedFriends components with styling updates
