+++ public/style.css
@@ -1023,6 +1023,30 @@ input[type="search"]::-webkit-search-cancel-button {
   top: 61px;
 }
 
+.left-2{
+  left: 0.5rem;
+}
+
+.left-4{
+  left: 1rem;
+}
+
+.-left-2{
+  left: -0.5rem;
+}
+
+.-left-3{
+  left: -0.75rem;
+}
+
+.-left-1{
+  left: -0.25rem;
+}
+
+.bottom-1{
+  bottom: 0.25rem;
+}
+
 .-z-10{
   z-index: -10;
 }
@@ -2890,6 +2914,26 @@ input[type="search"]::-webkit-search-cancel-button {
 
   .lg\:right-4{
     right: 1rem;
+  }
+
+  .lg\:left-4{
+    left: 1rem;
+  }
+
+  .lg\:-left-4{
+    left: -1rem;
+  }
+
+  .lg\:-left-5{
+    left: -1.25rem;
+  }
+
+  .lg\:-left-3{
+    left: -0.75rem;
+  }
+
+  .lg\:-left-3\.5{
+    left: -0.875rem;
   }
 
   .lg\:my-1{
